feat: searchDialog.tsx 분리
