+++ src/constants.ts
@@ -4,6 +4,7 @@ export const CellSize = {
   WIDTH: 80,
   HEIGHT: 30,
 };
+export const PAGE_SIZE = 20;
 
 export const 초 = 1000;
 export const 분 = 60 * 초;
